done Topic 26 (POM 09 - Check Undisplayed and Override global timeout)

+++ actions/commons/Constants.java
@@ -5,7 +5,7 @@ public class Constants {
 	public static final String TEST_APP_URL = "http://demo.guru99.com/v2";
 	public static final String STAGING_APP_URL = "http://demo.guru99.com/v3";
 	public static final String PRODUCTION_APP_URL = "http://demo.guru99.com/v4";
-	public static final int SHORT_TIMEOUT = 5;
-	public static final int LONG_TIMEOUT = 30;
+	public static final long SHORT_TIMEOUT = 5;
+	public static final long LONG_TIMEOUT = 30;
 
 }
